Move HeadTitle to components/common
