--- airbyte-webapp/src/components/HeadTitle/index.tsx
@@ -1,3 +1,0 @@
-import HeadTitle from "./HeadTitle";
-
-export default HeadTitle;
