+++ airbyte-webapp/src/components/common/HeadTitle/index.ts
@@ -1,0 +1,1 @@
+export * from "./HeadTitle";
