+++ airbyte-webapp/src/packages/cloud/views/auth/LoginPage/LoginPage.tsx
@@ -5,7 +5,7 @@ import { NavigateOptions, To, useNavigate } from "react-router-dom";
 import * as yup from "yup";
 
 import { LabeledInput, Link } from "components";
-import HeadTitle from "components/HeadTitle";
+import { HeadTitle } from "components/common/HeadTitle";
 import { Button } from "components/ui/Button";
 
 import { PageTrackingCodes, useTrackPage } from "hooks/services/Analytics";
