+++ airbyte-webapp/src/pages/SettingsPage/SettingsPage.tsx
@@ -2,7 +2,7 @@ import React, { Suspense } from "react";
 import { FormattedMessage } from "react-intl";
 import { Navigate, Route, Routes, useLocation, useNavigate } from "react-router-dom";
 
-import HeadTitle from "components/HeadTitle";
+import { HeadTitle } from "components/common/HeadTitle";
 import LoadingPage from "components/LoadingPage";
 import MainPageWithScroll from "components/MainPageWithScroll";
 import { PageHeader } from "components/ui/PageHeader";
