+++ airbyte-webapp/src/pages/SettingsPage/pages/AccountPage/AccountPage.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { FormattedMessage } from "react-intl";
 
-import HeadTitle from "components/HeadTitle";
+import { HeadTitle } from "components/common/HeadTitle";
 
 import useWorkspaceEditor from "pages/SettingsPage/components/useWorkspaceEditor";
 import { useCurrentWorkspace } from "services/workspaces/WorkspacesService";
